Exception handling - delete

+++ src/main/java/com/javastudy/course/services/UserService.java
@@ -2,8 +2,10 @@ package com.javastudy.course.services;
 
 import com.javastudy.course.entities.User;
 import com.javastudy.course.repositories.UserRepository;
+import com.javastudy.course.services.exceptions.DataBaseException;
 import com.javastudy.course.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,7 +31,17 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+
+        try {
+            if (!repository.existsById(id)) {
+                throw new ResourceNotFoundException(id);
+            }
+
+            repository.deleteById(id);
+
+        } catch (DataIntegrityViolationException e) {
+            throw new DataBaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj) {
